Restructure transaction code
